feat(chatRoom): Refactor chat/[groupName] to be similar to old Feed page and route. Delete Feed component and feed route. Add BorderStack and BoxWrapper to styles. Remove Feed from header and add Chat. Import MessageItem into chat group page.
